Move dbapi exceptions into seperate module
